# openapi/paths.yaml
openapi: 3.0.3
info:
  title: Swagger Shotflix
  description: |-
    This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io).

    _If you're looking for the Swagger 2.0/OAS 2.0 version of Petstore, then click [here](https://editor.swagger.io/?url=https://petstore.swagger.io/v2/swagger.yaml).

    Some useful links:
    - [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)
    - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
  version: 1.0.11
servers:
  - url: http://localhost:8000/api/
    description: Development Server
...
paths:
...
  /auth/refresh:
    get:
      tags:
        - auth
      summary: Yêu cần làm mới token
      description: Làm mới token
      operationId: refreshToken
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Đổi mật khẩu thành công
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessForgotPwdVerify"
        "401":
          description: Không tìm thấy mã OTP cho email này
        "403":
          description: User Not Registered
        "404":
          description: Mã OTP đã hết hạn hoặc không hợp lệ
        "500":
          description: Internal server error